+++ .secrets.baseline
@@ -3,7 +3,7 @@
     "files": "go.sum|^.secrets.baseline$",
     "lines": null
   },
-  "generated_at": "2024-05-15T13:37:47Z",
+  "generated_at": "2024-05-15T14:19:02Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -114,7 +114,7 @@
         "hashed_secret": "c7c6508b19455e3e8040e60e9833fbede92e5d8e",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 155,
+        "line_number": 162,
         "type": "Secret Keyword",
         "verified_result": null
       }
@@ -123,7 +123,7 @@
       {
         "hashed_secret": "7e6a3680012346b94b54731e13d8a9ffa3790645",
         "is_verified": false,
-        "line_number": 202,
+        "line_number": 203,
         "type": "Secret Keyword",
         "verified_result": null
       }
